# docs/file_service_api-docs.json
{"openapi":"3.0.1","info":{"title":"File service API","contact":{"name":"Potryvaev Alexandr","email":"Alexpotr@yandex.ru"}},"servers":[{"url":"http://localhost:8092","description":"Generated server url"}],"paths":{"/api/picture/uploadFile":{"post":{"tags":["image-controller"],"summary":"Контролер сохранения изображения","description":"Сохраняет в файловой системе изображение по данному ID","operationId":"uploadFile","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"Successfully retrieved Image"}}}},"/api/picture/recognise/{imageId}":{"post":{"tags":["image-controller"],"summary":"Контролер передачи изображения в сервис распознования","description":"Передает изображение по данному ID генерации midi файла в сервис распознования","operationId":"recognise","parameters":[{"name":"imageId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Image is not found"},"200":{"description":"Successfully send to download Image"}}}},"/api/midi/uploadFile":{"post":{"tags":["midi-controller"],"summary":"Контроллер загрузки ногвого midi файла","description":"Сохроняет в файловой системе данный midi файл","operationId":"uploadMidiFile","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"Successfully uploaded MIDI file"}}}},"/api/picture/{imageId}":{"get":{"tags":["image-controller"],"summary":"Контроллер просмотра изображения","description":"Передает изображение для просмотра веб-браузером пользователем по данному ID","operationId":"getImageById","parameters":[{"name":"imageId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully send Image"},"404":{"description":"Image is not found"}}}},"/api/picture/download/{imageId}":{"get":{"tags":["image-controller"],"summary":"Контроллер скачивания изображения","description":"Передает изображение для загрузки пользователем по данному ID","operationId":"sendImage","parameters":[{"name":"imageId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully send to download Image","content":{"application/octet-stream":{"schema":{"type":"object"}}}},"404":{"description":"Image is not found"}}}},"/api/midi/{midiId}":{"get":{"tags":["midi-controller"],"summary":"Контроллер получения midi файла","description":"Возвращает midi файл по данному ID","operationId":"getMidiById","parameters":[{"name":"midiId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"MIDI file not found"},"200":{"description":"Successfully retrieved MIDI file","content":{"audio/midi":{"schema":{"type":"array","items":{"type":"string","format":"byte"}}}}}}},"delete":{"tags":["midi-controller"],"summary":"Контроллер удаления midi файла","description":"Удаляет midi файл по данному ID","operationId":"deleteMidiById","parameters":[{"name":"midiId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully deleted MIDI file"},"404":{"description":"MIDI file is not found"}}}},"/api/picture/image/{imageId}":{"delete":{"tags":["image-controller"],"summary":"Контролер удаления изображения","description":"Удаляет в файловой системе изображение по данному ID","operationId":"deleteImage","parameters":[{"name":"imageId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully deleted image"},"404":{"description":"Image is not found"}}}}},"components":{}}